pyproject.toml
[project]
name = "auglab"
version = "20250905"
requires-python = ">=3.10"
description = "AugLab investigates the influence of different data augmentation strategies on MRI training performance."
readme = "README.md"
authors = [
    { name = "Nathan Molinier", email = "nathan.molinier@polymtl.ca"},
    { name = "Hendrik Möller"},
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
keywords = [
    'deep learning',
    'image segmentation',
    'nnU-Net',
    'nnunet',
    'magnetic resonance imaging',
    'online augmentation',
    'offline augmentation',
    'mri'
]
dependencies = [
    "batchgeneratorsv2",
    "nnunetv2",
    "kornia"
]

[project.optional-dependencies]
all = [
    "torchio",
    "monai[all]",
    "progress",
    "numpy",
    "tqdm",
    "wandb",
]

[project.scripts]
auglab_add_nnunettrainer = "auglab.add_trainer:main"

[project.urls]
homepage = "https://github.com/neuropoly/AugLab"
repository = "https://github.com/neuropoly/AugLab"

[build-system]
requires = ["pip>=23", "setuptools>=67"]
build-backend = "setuptools.build_meta"
